+++ src/imports/utils/cursorutils.ts
@@ -28,7 +28,7 @@ export function getObjectLinkedToAttribute(document: vscode.TextDocument, positi
 				const textBeforeBrace = textBeforeAttribute.substring(0, i).trim();
 
 				// Use a regex to find the object name before the '{'
-				const objectMatch = textBeforeBrace.match(/(?<= )([@~]|(\?~?!?))?\w+$/);  // Match the last word before the brace
+				const objectMatch = textBeforeBrace.match(/(?<= )([@~]|(\?~?!?))?[\w:]+$/);  // Match the last word before the brace
 				if (objectMatch && objectMatch[0]) {
 					return objectMatch[0];  // Return the object name
 				}